split style components
