+++ client/packages/lowcoder/src/components/assistant-ui/thread.tsx
@@ -1,418 +1,20 @@
-import {
-  ActionBarMorePrimitive,
-  ActionBarPrimitive,
-  AuiIf,
-  BranchPickerPrimitive,
-  ComposerPrimitive,
-  ErrorPrimitive,
-  getMcpAppFromToolPart,
-  MessagePrimitive,
-  ThreadPrimitive,
-  useAuiState,
-} from "@assistant-ui/react";
-import {
-  ArrowDownIcon,
-  ArrowUpIcon,
-  CheckIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  CopyIcon,
-  DownloadIcon,
-  MoreHorizontalIcon,
-  PencilIcon,
-  RefreshCwIcon,
-  SquareIcon,
-} from "lucide-react";
-import { type FC } from "react";
-import { Flex, Spin } from "antd";
+import { AuiIf, ThreadPrimitive } from "@assistant-ui/react";
 import { LoadingOutlined } from "@ant-design/icons";
-import styled from "styled-components";
+import { Flex, Spin } from "antd";
+import { ArrowDownIcon } from "lucide-react";
+import type { FC } from "react";
 import { trans } from "i18n";
 
-import {
-  ComposerAddAttachment,
-  ComposerAttachments,
-  UserMessageAttachments,
-} from "./ui/attachment";
-import { MarkdownText } from "./markdown-text";
-import {
-  Reasoning,
-  ReasoningContent,
-  ReasoningRoot,
-  ReasoningText,
-  ReasoningTrigger,
-} from "./reasoning";
-import {
-  ToolGroupContent,
-  ToolGroupRoot,
-  ToolGroupTrigger,
-} from "./tool-group";
-import { ToolFallback } from "./tool-fallback";
+import { Composer } from "./thread-composer";
+import { StyledThreadRoot } from "./thread.styles";
+import { ThreadMessage } from "./thread-message";
+import { ThreadWelcome } from "./thread-welcome";
 import { TooltipIconButton } from "./tooltip-icon-button";
-import { Button } from "./ui/button";
-
 
 interface ThreadProps {
   placeholder?: string;
   showAttachments?: boolean;
 }
-
-const StyledThreadRoot = styled(ThreadPrimitive.Root)`
-  background: #f9fafb;
-  color: #1f2937;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  min-height: 0;
-  overflow: hidden;
-
-  .aui-thread-viewport {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    min-height: 0;
-    overflow-x: hidden;
-    overflow-y: auto;
-    scroll-behavior: smooth;
-  }
-
-  .aui-thread-layout {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    margin: 0 auto;
-    max-width: var(--thread-max-width);
-    min-height: 0;
-    padding: 16px;
-    width: 100%;
-  }
-
-  .aui-message-group {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    margin-bottom: 24px;
-  }
-
-  .aui-message-group:empty {
-    display: none;
-  }
-
-  .aui-thread-welcome-root {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    justify-content: center;
-    margin: auto 0;
-  }
-
-  .aui-thread-welcome-center {
-    align-items: center;
-    display: flex;
-    flex: 1;
-    justify-content: center;
-    width: 100%;
-  }
-
-  .aui-thread-welcome-message {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding: 0 16px;
-    width: 100%;
-  }
-
-  .aui-thread-welcome-message-inner {
-    color: #111827;
-    font-size: 20px;
-    font-weight: 600;
-    line-height: 28px;
-    margin: 0;
-  }
-
-  .aui-thread-welcome-suggestions {
-    display: grid;
-    gap: 8px;
-    grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
-    padding-bottom: 16px;
-    width: 100%;
-  }
-
-  .aui-thread-welcome-suggestion {
-    align-items: flex-start;
-    background: #ffffff;
-    border: 1px solid #d9d9d9;
-    border-radius: 16px;
-    color: #1f2937;
-    cursor: pointer;
-    display: flex;
-    flex-wrap: wrap;
-    font-size: 14px;
-    gap: 4px;
-    min-height: 44px;
-    padding: 10px 14px;
-    text-align: left;
-    transition:
-      background-color 0.2s ease,
-      border-color 0.2s ease;
-    width: 100%;
-  }
-
-  .aui-thread-welcome-suggestion:hover {
-    background: #f3f4f6;
-    border-color: #bfbfbf;
-  }
-
-  .aui-thread-welcome-suggestion-text-1 {
-    font-weight: 500;
-  }
-
-  .aui-thread-viewport-footer {
-    background: linear-gradient(180deg, rgba(249, 250, 251, 0), #f9fafb 24px);
-    bottom: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    margin-top: auto;
-    overflow: visible;
-    padding-bottom: 16px;
-    padding-top: 24px;
-    position: sticky;
-    z-index: 2;
-  }
-
-  .aui-thread-scroll-to-bottom {
-    align-self: center;
-    box-shadow: 0 4px 12px rgba(15, 23, 42, 0.12);
-  }
-
-  .aui-composer-root {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    width: 100%;
-  }
-
-  .aui-composer-shell {
-    background: #ffffff;
-    border: 1px solid #d9d9d9;
-    border-radius: var(--composer-radius);
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    padding: var(--composer-padding);
-    transition:
-      border-color 0.2s ease,
-      box-shadow 0.2s ease;
-    width: 100%;
-  }
-
-  .aui-composer-shell:focus-within {
-    border-color: #1677ff;
-    box-shadow: 0 0 0 2px rgba(22, 119, 255, 0.12);
-  }
-
-  .aui-composer-input {
-    background: transparent;
-    border: 0;
-    color: #1f2937;
-    font-size: 14px;
-    line-height: 22px;
-    max-height: 128px;
-    min-height: 40px;
-    outline: 0;
-    padding: 4px 6px;
-    resize: none;
-    width: 100%;
-  }
-
-  .aui-composer-input::placeholder {
-    color: #8c8c8c;
-  }
-
-  .aui-composer-action-wrapper {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    min-height: 32px;
-  }
-
-  .aui-composer-add-attachment,
-  .aui-composer-send,
-  .aui-composer-cancel {
-    border-radius: 50%;
-  }
-
-  .aui-assistant-message-root {
-    margin: 0 auto;
-    max-width: var(--thread-max-width);
-    position: relative;
-    width: 100%;
-  }
-
-  .aui-assistant-message-content {
-    color: #1f2937;
-    font-size: 14px;
-    line-height: 24px;
-    padding: 0 8px;
-    word-break: break-word;
-  }
-
-  .aui-assistant-message-footer {
-    align-items: center;
-    display: flex;
-    margin-left: 8px;
-    min-height: 28px;
-  }
-
-  .aui-assistant-action-bar-root {
-    align-items: center;
-    display: flex;
-    gap: 4px;
-    color: #6b7280;
-  }
-
-  .aui-action-bar-more-content {
-    background: #ffffff;
-    border: 1px solid #e5e7eb;
-    border-radius: 8px;
-    box-shadow: 0 8px 24px rgba(15, 23, 42, 0.12);
-    min-width: 144px;
-    padding: 4px;
-    z-index: 1000;
-  }
-
-  .aui-action-bar-more-item {
-    align-items: center;
-    border-radius: 6px;
-    color: #1f2937;
-    cursor: pointer;
-    display: flex;
-    font-size: 14px;
-    gap: 8px;
-    outline: none;
-    padding: 7px 8px;
-  }
-
-  .aui-action-bar-more-item:hover,
-  .aui-action-bar-more-item:focus {
-    background: #f3f4f6;
-  }
-
-  .aui-user-message-root {
-    animation: none;
-    display: grid;
-    grid-template-columns: minmax(72px, 1fr) auto;
-    grid-auto-rows: auto;
-    gap: 8px 0;
-    margin: 0 auto;
-    max-width: var(--thread-max-width);
-    padding: 0 8px;
-    width: 100%;
-  }
-
-  .aui-user-message-content-wrapper {
-    grid-column-start: 2;
-    min-width: 0;
-    position: relative;
-  }
-
-  .aui-user-message-content {
-    background: #e5e7eb;
-    border-radius: 16px;
-    color: #111827;
-    font-size: 14px;
-    line-height: 22px;
-    padding: 10px 16px;
-    word-break: break-word;
-  }
-
-  .aui-user-action-bar-wrapper {
-    padding-right: 8px;
-    position: absolute;
-    right: 100%;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-
-  .aui-user-action-bar-root {
-    align-items: flex-end;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .aui-user-branch-picker {
-    grid-column: 1 / -1;
-    grid-row-start: 3;
-    justify-content: flex-end;
-  }
-
-  .aui-edit-composer-wrapper {
-    display: flex;
-    flex-direction: column;
-    padding: 0 8px;
-  }
-
-  .aui-edit-composer-root {
-    align-self: flex-end;
-    background: #e5e7eb;
-    border-radius: 16px;
-    display: flex;
-    flex-direction: column;
-    max-width: 85%;
-    width: 100%;
-  }
-
-  .aui-edit-composer-input {
-    background: transparent;
-    border: 0;
-    color: #111827;
-    font-size: 14px;
-    min-height: 56px;
-    outline: 0;
-    padding: 16px;
-    resize: none;
-    width: 100%;
-  }
-
-  .aui-edit-composer-footer {
-    align-items: center;
-    align-self: flex-end;
-    display: flex;
-    gap: 8px;
-    margin: 0 12px 12px;
-  }
-
-  .aui-branch-picker-root {
-    align-items: center;
-    color: #6b7280;
-    display: inline-flex;
-    font-size: 12px;
-    gap: 4px;
-    margin-left: -8px;
-    margin-right: 8px;
-  }
-
-  .aui-branch-picker-state {
-    font-weight: 500;
-    min-width: 36px;
-    text-align: center;
-  }
-
-  .aui-tooltip-icon-button svg,
-  .aui-button svg {
-    height: 16px;
-    width: 16px;
-  }
-
-  .aui-assistant-message-root:has([data-status="running"]) {
-    display: none;
-  }
-
-  .aui-assistant-message-content [data-status="running"] {
-    display: none;
-  }
-`;
 
 const SimpleANTDLoader = () => {
   const antIcon = <LoadingOutlined style={{ fontSize: 24 }} spin />;
@@ -452,10 +54,7 @@ export const Thread: FC<ThreadProps> = ({
             <ThreadWelcome />
           </AuiIf>
 
-          <div
-            data-slot="aui_message-group"
-            className="aui-message-group"
-          >
+          <div data-slot="aui_message-group" className="aui-message-group">
             <ThreadPrimitive.Messages>
               {() => <ThreadMessage showAttachments={showAttachments} />}
             </ThreadPrimitive.Messages>
@@ -478,17 +77,6 @@ export const Thread: FC<ThreadProps> = ({
   );
 };
 
-const ThreadMessage: FC<{ showAttachments?: boolean }> = ({
-  showAttachments = true,
-}) => {
-  const role = useAuiState((s) => s.message.role);
-  const isEditing = useAuiState((s) => s.message.composer.isEditing);
-
-  if (isEditing) return <EditComposer />;
-  if (role === "user") return <UserMessage showAttachments={showAttachments} />;
-  return <AssistantMessage />;
-};
-
 const ThreadScrollToBottom: FC = () => {
   return (
     <ThreadPrimitive.ScrollToBottom asChild>
@@ -500,340 +88,5 @@ const ThreadScrollToBottom: FC = () => {
         <ArrowDownIcon />
       </TooltipIconButton>
     </ThreadPrimitive.ScrollToBottom>
-  );
-};
-
-const ThreadWelcome: FC = () => {
-  return (
-    <div className="aui-thread-welcome-root">
-      <div className="aui-thread-welcome-center">
-        <div className="aui-thread-welcome-message">
-          <h1 className="aui-thread-welcome-message-inner">
-            {trans("chat.welcomeMessage")}
-          </h1>
-        </div>
-      </div>
-      <ThreadSuggestions />
-    </div>
-  );
-};
-
-const ThreadSuggestions: FC = () => {
-  return (
-    <div className="aui-thread-welcome-suggestions">
-      <div className="aui-thread-welcome-suggestion-display">
-        <ThreadPrimitive.Suggestion
-          className="aui-thread-welcome-suggestion"
-          prompt={trans("chat.suggestionWeather")}
-          method="replace"
-          autoSend
-        >
-          <span className="aui-thread-welcome-suggestion-text-1">
-            {trans("chat.suggestionWeather")}
-          </span>
-        </ThreadPrimitive.Suggestion>
-      </div>
-      <div className="aui-thread-welcome-suggestion-display">
-        <ThreadPrimitive.Suggestion
-          className="aui-thread-welcome-suggestion"
-          prompt={trans("chat.suggestionAssistant")}
-          method="replace"
-          autoSend
-        >
-          <span className="aui-thread-welcome-suggestion-text-1">
-            {trans("chat.suggestionAssistant")}
-          </span>
-        </ThreadPrimitive.Suggestion>
-      </div>
-    </div>
-  );
-};
-
-const Composer: FC<{ placeholder?: string; showAttachments?: boolean }> = ({
-  placeholder = trans("chat.composerPlaceholder"),
-  showAttachments = true,
-}) => {
-  return (
-    <ComposerPrimitive.Root className="aui-composer-root">
-      <ComposerPrimitive.AttachmentDropzone asChild>
-        <div
-          data-slot="aui_composer-shell"
-          className="aui-composer-shell"
-        >
-          {showAttachments && <ComposerAttachments />}
-          <ComposerPrimitive.Input
-            placeholder={placeholder}
-            className="aui-composer-input"
-            rows={1}
-            autoFocus
-            aria-label="Message input"
-          />
-          <ComposerAction showAttachments={showAttachments} />
-        </div>
-      </ComposerPrimitive.AttachmentDropzone>
-    </ComposerPrimitive.Root>
-  );
-};
-
-const ComposerAction: FC<{ showAttachments?: boolean }> = ({
-  showAttachments = true,
-}) => {
-  return (
-    <div className="aui-composer-action-wrapper">
-      {showAttachments ? <ComposerAddAttachment /> : <span />}
-      <AuiIf condition={(s) => !s.thread.isRunning}>
-        <ComposerPrimitive.Send asChild>
-          <TooltipIconButton
-            tooltip="Send message"
-            side="bottom"
-            type="button"
-            variant="default"
-            size="icon"
-            className="aui-composer-send"
-            aria-label="Send message"
-          >
-            <ArrowUpIcon className="aui-composer-send-icon" />
-          </TooltipIconButton>
-        </ComposerPrimitive.Send>
-      </AuiIf>
-      <AuiIf condition={(s) => s.thread.isRunning}>
-        <ComposerPrimitive.Cancel asChild>
-          <Button
-            type="button"
-            variant="default"
-            size="icon"
-            className="aui-composer-cancel"
-            aria-label="Stop generating"
-          >
-            <SquareIcon className="aui-composer-cancel-icon" />
-          </Button>
-        </ComposerPrimitive.Cancel>
-      </AuiIf>
-    </div>
-  );
-};
-
-const MessageError: FC = () => {
-  return (
-    <MessagePrimitive.Error>
-      <ErrorPrimitive.Root className="aui-message-error-root">
-        <ErrorPrimitive.Message className="aui-message-error-message" />
-      </ErrorPrimitive.Root>
-    </MessagePrimitive.Error>
-  );
-};
-
-const AssistantMessage: FC = () => {
-  return (
-    <MessagePrimitive.Root
-      data-slot="aui_assistant-message-root"
-      data-role="assistant"
-      className="aui-assistant-message-root"
-    >
-      <div
-        data-slot="aui_assistant-message-content"
-        className="aui-assistant-message-content"
-      >
-        <MessagePrimitive.GroupedParts
-          groupBy={(part) => {
-            if (part.type === "reasoning")
-              return ["group-chainOfThought", "group-reasoning"];
-            if (part.type === "tool-call") {
-              if (getMcpAppFromToolPart(part)) return null;
-              return ["group-chainOfThought", "group-tool"];
-            }
-            return null;
-          }}
-        >
-          {({ part, children }) => {
-            switch (part.type) {
-              case "group-chainOfThought":
-                return <div data-slot="aui_chain-of-thought">{children}</div>;
-              case "group-reasoning": {
-                const running = part.status.type === "running";
-                return (
-                  <ReasoningRoot defaultOpen={running}>
-                    <ReasoningTrigger active={running} />
-                    <ReasoningContent aria-busy={running}>
-                      <ReasoningText>{children}</ReasoningText>
-                    </ReasoningContent>
-                  </ReasoningRoot>
-                );
-              }
-              case "group-tool":
-                return (
-                  <ToolGroupRoot>
-                    <ToolGroupTrigger
-                      count={part.indices.length}
-                      active={part.status.type === "running"}
-                    />
-                    <ToolGroupContent>{children}</ToolGroupContent>
-                  </ToolGroupRoot>
-                );
-              case "text":
-                return <MarkdownText />;
-              case "reasoning":
-                return <Reasoning {...part} />;
-              case "tool-call":
-                return part.toolUI ?? <ToolFallback {...part} />;
-              default:
-                return null;
-            }
-          }}
-        </MessagePrimitive.GroupedParts>
-        <MessageError />
-      </div>
-
-      <div
-        data-slot="aui_assistant-message-footer"
-        className="aui-assistant-message-footer"
-      >
-        <BranchPicker />
-        <AssistantActionBar />
-      </div>
-    </MessagePrimitive.Root>
-  );
-};
-
-const AssistantActionBar: FC = () => {
-  return (
-    <ActionBarPrimitive.Root
-      hideWhenRunning
-      autohide="not-last"
-      className="aui-assistant-action-bar-root"
-    >
-      <ActionBarPrimitive.Copy asChild>
-        <TooltipIconButton tooltip="Copy">
-          <AuiIf condition={(s) => s.message.isCopied}>
-            <CheckIcon />
-          </AuiIf>
-          <AuiIf condition={(s) => !s.message.isCopied}>
-            <CopyIcon />
-          </AuiIf>
-        </TooltipIconButton>
-      </ActionBarPrimitive.Copy>
-      <ActionBarPrimitive.Reload asChild>
-        <TooltipIconButton tooltip="Refresh">
-          <RefreshCwIcon />
-        </TooltipIconButton>
-      </ActionBarPrimitive.Reload>
-      <ActionBarMorePrimitive.Root>
-        <ActionBarMorePrimitive.Trigger asChild>
-          <TooltipIconButton
-            tooltip="More"
-            className="aui-action-bar-more-trigger"
-          >
-            <MoreHorizontalIcon />
-          </TooltipIconButton>
-        </ActionBarMorePrimitive.Trigger>
-        <ActionBarMorePrimitive.Content
-          side="bottom"
-          align="start"
-          className="aui-action-bar-more-content"
-        >
-          <ActionBarPrimitive.ExportMarkdown asChild>
-            <ActionBarMorePrimitive.Item className="aui-action-bar-more-item">
-              <DownloadIcon />
-              Export as Markdown
-            </ActionBarMorePrimitive.Item>
-          </ActionBarPrimitive.ExportMarkdown>
-        </ActionBarMorePrimitive.Content>
-      </ActionBarMorePrimitive.Root>
-    </ActionBarPrimitive.Root>
-  );
-};
-
-const UserMessage: FC<{ showAttachments?: boolean }> = ({
-  showAttachments = true,
-}) => {
-  return (
-    <MessagePrimitive.Root
-      data-slot="aui_user-message-root"
-      className="aui-user-message-root"
-      data-role="user"
-    >
-      {showAttachments && <UserMessageAttachments />}
-
-      <div className="aui-user-message-content-wrapper">
-        <div className="aui-user-message-content">
-          <MessagePrimitive.Parts />
-        </div>
-        <div className="aui-user-action-bar-wrapper">
-          <UserActionBar />
-        </div>
-      </div>
-
-      <BranchPicker
-        data-slot="aui_user-branch-picker"
-        className="aui-user-branch-picker"
-      />
-    </MessagePrimitive.Root>
-  );
-};
-
-const UserActionBar: FC = () => {
-  return (
-    <ActionBarPrimitive.Root
-      hideWhenRunning
-      autohide="not-last"
-      className="aui-user-action-bar-root"
-    >
-      <ActionBarPrimitive.Edit asChild>
-        <TooltipIconButton tooltip="Edit" className="aui-user-action-edit">
-          <PencilIcon />
-        </TooltipIconButton>
-      </ActionBarPrimitive.Edit>
-    </ActionBarPrimitive.Root>
-  );
-};
-
-const EditComposer: FC = () => {
-  return (
-    <MessagePrimitive.Root
-      data-slot="aui_edit-composer-wrapper"
-      className="aui-edit-composer-wrapper"
-    >
-      <ComposerPrimitive.Root className="aui-edit-composer-root">
-        <ComposerPrimitive.Input
-          className="aui-edit-composer-input"
-          autoFocus
-        />
-        <div className="aui-edit-composer-footer">
-          <ComposerPrimitive.Cancel asChild>
-            <Button variant="ghost">Cancel</Button>
-          </ComposerPrimitive.Cancel>
-          <ComposerPrimitive.Send asChild>
-            <Button>Update</Button>
-          </ComposerPrimitive.Send>
-        </div>
-      </ComposerPrimitive.Root>
-    </MessagePrimitive.Root>
-  );
-};
-
-const BranchPicker: FC<BranchPickerPrimitive.Root.Props> = ({
-  className,
-  ...rest
-}) => {
-  return (
-    <BranchPickerPrimitive.Root
-      hideWhenSingleBranch
-      className={className ? `aui-branch-picker-root ${className}` : "aui-branch-picker-root"}
-      {...rest}
-    >
-      <BranchPickerPrimitive.Previous asChild>
-        <TooltipIconButton tooltip="Previous">
-          <ChevronLeftIcon />
-        </TooltipIconButton>
-      </BranchPickerPrimitive.Previous>
-      <span className="aui-branch-picker-state">
-        <BranchPickerPrimitive.Number /> / <BranchPickerPrimitive.Count />
-      </span>
-      <BranchPickerPrimitive.Next asChild>
-        <TooltipIconButton tooltip="Next">
-          <ChevronRightIcon />
-        </TooltipIconButton>
-      </BranchPickerPrimitive.Next>
-    </BranchPickerPrimitive.Root>
   );
 };
